fix: resolve CodeQL circular import between job_queue and jobs.queue.config
Moves _get_redis_settings from job_queue.py to jobs/queue/config.py
to break the import cycle: job_queue -> jobs.queue.config -> job_queue.
The function is re-exported from job_queue.py so all existing importers
continue to work unchanged.

CodeQL finding (4 errors in check-run 81616620946):
- WorkerSettings may not be defined due to cyclic import
- _worker_on_startup, _worker_cron_jobs, _worker_redis_settings same
